requirements.txt
numpy==1.14.2
Pillow==5.0.0
pip==9.0.1
pygame==1.9.3
PyOpenGL==3.1.0
setuptools==28.8.0
keras>=2.0.5
opencv-python>=2.4.9
tensorflow>=1.6.0
scikit-learn>=0.19.1
